Charts!
Swapped out the Wpf Toolkit DataVisulisations as they were really slow and replaced them with amCharts.
Put locks around the code which reads and writes from the PingHistory as this is a volitile collection. So we prevent errors being thrown by the UI and the random dips in the graph for the Average line where an error was thrown.